+++ presto-main/src/main/java/com/facebook/presto/testing/TestngUtils.java
@@ -20,9 +20,9 @@ import java.util.stream.Stream;
 
 import static java.util.Arrays.asList;
 
-public class Arguments
+public final class TestngUtils
 {
-    private Arguments()
+    private TestngUtils()
     {
     }
